refactor(practice-process): relocate exception constants to a separate file
